backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
django-redis==5.4.0
django-axes==6.1.1
weaviate-client==3.25.2
python-dotenv==1.0.0
Pillow==10.1.0
pdfplumber==0.10.3
sentence-transformers==2.2.2
openai==1.12.0
transformers==4.38.1
pytest==7.4.3
pytest-django==4.7.0
black==23.11.0
isort==5.12.0
django-cors-headers==4.3.1
gunicorn==21.2.0
dj-database-url==2.1.0